Fix direct lookup on empty list
When Vale has no results to display, the `temp_list` variable should be
empty. However, we need to check for config run time errors, which
appear in the first element of the list.

This commit adds a length-checker to verify that the list has one
element or more prior performing run-time error checking.