Simplify Vale output generation
With newer versions of Vale, the workaround of creating a temp .md file
to chunk in all the changes isn't necessary. It's simpler to just run
the command and parse it into the quickfix list.

This commit revamps a few things and gets rid of the whole temp_file.lua
module as consequence.